restore DataSource config for development.
I don't have the external datasource.groovy, so I restored an earlier
DataSource.groovy to use locally.  The passwords were already removed
from the git history, and should remain external, but the local,
in-memory hsqldb driver password does not need to be secret.

This config is just to quickly test on my Mac, not a permanent suggestion.