Remove trailing / on cas.serverUrlPrefix.
This is for compatibility with CAS 6.3.

Also enable external config files, and move environment-specific
CAS config into the appropriate environments closure, instead of
commenting and uncommenting.

We are having (dependency) problems rebuilding the war,
and a tight deadline for upgrading CAS to 6.3, so
we're going to deploy just these Config*.class files, by manually
replacing them in a copy of the war file that is currently deployed
in production, and then deploying that modified war file.